fix: address import ingress review comments
